+++ src/commands/tx.rs
@@ -7,6 +7,7 @@ use crate::types::{
     RegisterCollectionData, RegisterOwnershipData, SlotRanges,
 };
 use crate::wallet::passphrase::prompt_passphrase_once;
+use crate::wallet::utxo_selection;
 use crate::{cli, context, wallet::brc721_wallet::Brc721Wallet};
 use age::secrecy::SecretString;
 use anyhow::{anyhow, Context, Result};
@@ -90,8 +91,10 @@ fn run_register_collection(
     if db_path.exists() {
         let storage = crate::storage::SqliteStorage::new(&db_path);
         let wallet_utxos = wallet.list_unspent(0).context("list wallet UTXOs")?;
-        lock_outpoints = compute_wallet_token_outpoints_to_lock(&storage, &wallet_utxos, &[])
-            .context("compute lock set")?;
+        let ownership_outpoints =
+            utxo_selection::ownership_outpoints_for_wallet(&storage, &wallet_utxos)
+                .context("compute ownership outpoints")?;
+        lock_outpoints = utxo_selection::lock_outpoints_for_fees(&ownership_outpoints, &[]);
     } else {
         log::warn!(
             "scanner database not found at {} (proceeding without ownership UTXO locks)",
@@ -135,11 +138,14 @@ fn run_register_ownership(
     let mut wallet = load_wallet(ctx)?;
     let wallet_utxos = wallet.list_unspent(0).context("list wallet UTXOs")?;
     let mut lock_outpoints = Vec::new();
+    let mut ownership_outpoints = BTreeSet::new();
     let db_path = ctx.data_dir.join("brc721.sqlite");
     if db_path.exists() {
         let storage = crate::storage::SqliteStorage::new(&db_path);
-        lock_outpoints = compute_wallet_token_outpoints_to_lock(&storage, &wallet_utxos, &[])
-            .context("compute lock set")?;
+        ownership_outpoints =
+            utxo_selection::ownership_outpoints_for_wallet(&storage, &wallet_utxos)
+                .context("compute ownership outpoints")?;
+        lock_outpoints = utxo_selection::lock_outpoints_for_fees(&ownership_outpoints, &[]);
     } else if init_owner.is_some() {
         return Err(anyhow!(
             "scanner database not found at {} (required when using --init-owner)",
@@ -189,12 +195,11 @@ fn run_register_ownership(
             }
         },
     };
-    let lock_set = lock_outpoints.iter().cloned().collect::<BTreeSet<_>>();
     let mandatory_inputs = match &init_spec {
-        Some(spec) => vec![select_init_owner_outpoint(
+        Some(spec) => vec![utxo_selection::select_non_ownership_utxo_for_address(
             &wallet_utxos,
             &spec.address,
-            &lock_set,
+            &ownership_outpoints,
         )?],
         None => Vec::new(),
     };
@@ -246,8 +251,10 @@ fn run_send_amount(
     if db_path.exists() {
         let storage = crate::storage::SqliteStorage::new(&db_path);
         let wallet_utxos = wallet.list_unspent(0).context("list wallet UTXOs")?;
-        lock_outpoints = compute_wallet_token_outpoints_to_lock(&storage, &wallet_utxos, &[])
-            .context("compute lock set")?;
+        let ownership_outpoints =
+            utxo_selection::ownership_outpoints_for_wallet(&storage, &wallet_utxos)
+                .context("compute ownership outpoints")?;
+        lock_outpoints = utxo_selection::lock_outpoints_for_fees(&ownership_outpoints, &[]);
     } else {
         log::warn!(
             "scanner database not found at {} (proceeding without ownership UTXO locks)",
@@ -335,9 +342,11 @@ fn run_send_assets(
     let dust_amount = Amount::from_sat(dust_sat);
     let passphrase = resolve_passphrase(passphrase)?;
 
+    let ownership_outpoints =
+        utxo_selection::ownership_outpoints_for_wallet(&storage, &wallet_utxos)
+            .context("compute ownership outpoints")?;
     let lock_outpoints =
-        compute_wallet_token_outpoints_to_lock(&storage, &wallet_utxos, &token_outpoints)
-            .context("compute lock set")?;
+        utxo_selection::lock_outpoints_for_fees(&ownership_outpoints, &token_outpoints);
 
     let tx = wallet
         .build_implicit_transfer_tx(
@@ -455,9 +464,11 @@ fn run_mix(
         }
     }
 
+    let ownership_outpoints =
+        utxo_selection::ownership_outpoints_for_wallet(&storage, &wallet_utxos)
+            .context("compute ownership outpoints")?;
     let lock_outpoints =
-        compute_wallet_token_outpoints_to_lock(&storage, &wallet_utxos, &token_outpoints)
-            .context("compute lock set")?;
+        utxo_selection::lock_outpoints_for_fees(&ownership_outpoints, &token_outpoints);
 
     let dust_amount = Amount::from_sat(dust_sat);
     let payments = output_addresses
@@ -556,48 +567,6 @@ fn resolve_owner_spec(
     })
 }
 
-fn select_init_owner_outpoint(
-    wallet_utxos: &[bitcoincore_rpc::json::ListUnspentResultEntry],
-    owner_address: &Address,
-    disallowed: &BTreeSet<OutPoint>,
-) -> Result<OutPoint> {
-    let script_pubkey = owner_address.script_pubkey();
-    let mut best: Option<&bitcoincore_rpc::json::ListUnspentResultEntry> = None;
-
-    for utxo in wallet_utxos {
-        if utxo.script_pub_key != script_pubkey {
-            continue;
-        }
-        let outpoint = OutPoint {
-            txid: utxo.txid,
-            vout: utxo.vout,
-        };
-        if disallowed.contains(&outpoint) {
-            continue;
-        }
-        match best {
-            None => best = Some(utxo),
-            Some(current) => {
-                if utxo.amount.to_sat() > current.amount.to_sat() {
-                    best = Some(utxo);
-                }
-            }
-        }
-    }
-
-    let Some(utxo) = best else {
-        return Err(anyhow!(
-            "no spendable UTXO found for init-owner address {} (fund it with a non-NFT UTXO)",
-            owner_address
-        ));
-    };
-
-    Ok(OutPoint {
-        txid: utxo.txid,
-        vout: utxo.vout,
-    })
-}
-
 fn parse_address_h160(raw: &str, flag_name: &str) -> Result<H160> {
     let trimmed = raw.trim();
     let trimmed = trimmed.strip_prefix("addressH160=").unwrap_or(trimmed);
@@ -622,38 +591,6 @@ fn parse_outpoints(outpoints: &[String]) -> Result<Vec<OutPoint>> {
                 .with_context(|| format!("invalid outpoint '{outpoint}' (expected TXID:VOUT)"))
         })
         .collect()
-}
-
-fn compute_wallet_token_outpoints_to_lock(
-    storage: &crate::storage::SqliteStorage,
-    wallet_utxos: &[bitcoincore_rpc::json::ListUnspentResultEntry],
-    spending: &[OutPoint],
-) -> Result<Vec<OutPoint>> {
-    let spending_set = spending.iter().cloned().collect::<BTreeSet<_>>();
-
-    // Build a set of wallet-owned outpoints that the index considers ownership UTXOs.
-    let mut wallet_token_outpoints = BTreeSet::new();
-    for utxo in wallet_utxos {
-        let txid = utxo.txid.to_string();
-        let vout = utxo.vout;
-        if storage
-            .list_unspent_ownership_utxos_by_outpoint(&txid, vout)
-            .with_context(|| format!("query ownership ranges for {txid}:{vout}"))?
-            .is_empty()
-        {
-            continue;
-        }
-
-        wallet_token_outpoints.insert(OutPoint {
-            txid: utxo.txid,
-            vout,
-        });
-    }
-
-    Ok(wallet_token_outpoints
-        .difference(&spending_set)
-        .cloned()
-        .collect())
 }
 
 fn parse_mix_outputs(
@@ -710,76 +647,10 @@ fn parse_mix_outputs(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use bitcoin::{Amount, Network, Txid};
-    use bitcoincore_rpc::json::ListUnspentResultEntry;
 
     #[test]
     fn parse_outpoints_rejects_invalid() {
         let res = parse_outpoints(&["not-an-outpoint".to_string()]);
-        assert!(res.is_err());
-    }
-
-    fn make_utxo(
-        txid_hex: &str,
-        vout: u32,
-        script_pub_key: bitcoin::ScriptBuf,
-        amount_sat: u64,
-    ) -> ListUnspentResultEntry {
-        ListUnspentResultEntry {
-            txid: Txid::from_str(txid_hex).unwrap(),
-            vout,
-            address: None,
-            label: None,
-            redeem_script: None,
-            witness_script: None,
-            script_pub_key,
-            amount: Amount::from_sat(amount_sat),
-            confirmations: 1,
-            spendable: true,
-            solvable: true,
-            descriptor: None,
-            safe: true,
-        }
-    }
-
-    #[test]
-    fn select_init_owner_outpoint_skips_disallowed() {
-        let address = Address::from_str("bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh")
-            .unwrap()
-            .require_network(Network::Bitcoin)
-            .unwrap();
-        let script_pub_key = address.script_pubkey();
-        let utxo_nft = make_utxo(&"00".repeat(32), 0, script_pub_key.clone(), 1000);
-        let utxo_ok = make_utxo(&"11".repeat(32), 1, script_pub_key.clone(), 2000);
-
-        let mut disallowed = BTreeSet::new();
-        disallowed.insert(OutPoint {
-            txid: utxo_nft.txid,
-            vout: utxo_nft.vout,
-        });
-
-        let selected =
-            select_init_owner_outpoint(&[utxo_nft, utxo_ok], &address, &disallowed).unwrap();
-        assert_eq!(selected.txid, Txid::from_str(&"11".repeat(32)).unwrap());
-        assert_eq!(selected.vout, 1);
-    }
-
-    #[test]
-    fn select_init_owner_outpoint_errors_when_only_disallowed() {
-        let address = Address::from_str("bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh")
-            .unwrap()
-            .require_network(Network::Bitcoin)
-            .unwrap();
-        let script_pub_key = address.script_pubkey();
-        let utxo_nft = make_utxo(&"00".repeat(32), 0, script_pub_key.clone(), 1000);
-
-        let mut disallowed = BTreeSet::new();
-        disallowed.insert(OutPoint {
-            txid: utxo_nft.txid,
-            vout: utxo_nft.vout,
-        });
-
-        let res = select_init_owner_outpoint(&[utxo_nft], &address, &disallowed);
         assert!(res.is_err());
     }
 }
